cloud_functions/api/get-dashboard-data/requirements.txt
functions-framework
firebase-admin
google-cloud-bigquery
Flask
